requirements.txt
# Core ML Dependencies
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
torch-geometric>=2.3.0
torchvision>=0.15.0

# Gradient Boosting
lightgbm>=4.0.0
xgboost>=1.7.0

# Deep Learning Utilities
timm>=0.9.0
einops>=0.6.0

# FastAPI & API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Async Tasks
celery[redis]>=5.3.0
redis>=4.5.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
alembic>=1.11.0

# UI Frameworks
gradio>=3.35.0
streamlit>=1.24.0
plotly>=5.14.0
dash>=2.11.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.8.0

# Web Scraping & Screenshots
selenium>=4.10.0
playwright>=1.35.0
pillow>=9.5.0

# Binary Analysis
pefile>=2023.2.7
lief>=0.13.0
networkx>=3.1

# MLOps
mlflow>=2.4.0
wandb>=0.15.0
prometheus-client>=0.17.0

# NLP
nltk>=3.8.0
spacy>=3.5.0

# Explainability
shap>=0.42.0
captum>=0.6.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
locust>=2.15.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0

# Monitoring
sentry-sdk>=1.25.0
python-json-logger>=2.0.7

# Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Production
gunicorn>=20.1.0
gevent>=23.7.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.4.0
